use argh because it doesn't require exceptions
