+++ runners/s3-benchrunner-c/CMakeLists.txt
@@ -2,9 +2,6 @@ cmake_minimum_required(VERSION 3.22)
 cmake_policy(VERSION 3.22...3.24)
 project(s3-benchrunner-c CXX)
 
-message(STATUS "C++ standard: ${CMAKE_CXX_STANDARD}")
-message(STATUS "C++ compiler flags: ${CMAKE_CXX_FLAGS}")
-
 add_executable(${PROJECT_NAME}
     BenchmarkRunner.cpp
     CRunner.cpp)
@@ -16,7 +13,7 @@ set_target_properties(${PROJECT_NAME} PROPERTIES
     CXX_STANDARD 20)
 
 if(NOT MSVC)
-    target_compile_options(${PROJECT_NAME} PRIVATE "-fno-rtti")
+    target_compile_options(${PROJECT_NAME} PRIVATE "-fno-exceptions;-fno-rtti")
 endif()
 
 # dependencies
@@ -26,18 +23,15 @@ include(FetchContent)
 FetchContent_Declare(json URL https://github.com/nlohmann/json/releases/download/v3.11.2/json.tar.xz)
 FetchContent_MakeAvailable(json)
 
-# using CLI11 for CLI argument parsing
-
-
+# CLI parser
 FetchContent_Declare(
-    cli11_proj
+    argh_parser
     QUIET
-    GIT_REPOSITORY https://github.com/CLIUtils/CLI11.git
-    GIT_TAG v2.5.0
+    GIT_REPOSITORY https://github.com/adishavit/argh
+    GIT_TAG v1.3.2
 )
-FetchContent_MakeAvailable(cli11_proj)
+FetchContent_MakeAvailable(argh_parser)
 
 find_package(aws-c-s3 REQUIRED)
 
-target_link_libraries(${PROJECT_NAME} PRIVATE AWS::aws-c-s3 nlohmann_json)
-target_link_libraries(${PROJECT_NAME} PRIVATE CLI11::CLI11)
+target_link_libraries(${PROJECT_NAME} PRIVATE AWS::aws-c-s3 nlohmann_json argh)
